% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_form.R
\name{delete_form}
\alias{delete_form}
\title{Deletes a typeform}
\usage{
delete_form(form_id, api = NULL)
}
\arguments{
\item{form_id}{The form id of the typeform you want the results for.}

\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api2")} is used.}
}
\description{
Delete the specified typeform.
}
