% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_responses.R
\name{delete_responses}
\alias{delete_responses}
\title{Delete Responses}
\usage{
delete_responses(form_id, api = NULL, included_tokens = NULL)
}
\arguments{
\item{form_id}{The form id of the typeform you want the results for.}

\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api2")} is used.}

\item{included_tokens}{A vector of tokens to delete. Maximum 1000. Token ids
can be extracted from the meta data frame via \code{get_form}.}
}
\value{
If successful, the function will return (invisibily) \code{TRUE}
}
\description{
Delete entries in a type form- use with care!
}
