% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_typeforms.R
\name{get_all_typeforms}
\alias{get_all_typeforms}
\title{Fetch all available typeforms}
\usage{
get_all_typeforms(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL}
we use the environment variable \code{Sys.getenv("typeform_api")}.}
}
\value{
A two column data frame.
}
\description{
This function returns a data frame containing your typeforms and their
associated UIDs.
}
\examples{
\dontrun{
api = "XXXXX"
get_all_typeforms(api)
}
}

