% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_api}
\alias{get_api}
\title{Retrieve API key from Renviron}
\usage{
get_api(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL}
we use the environment variable \code{Sys.getenv("typeform_api")}.}
}
\description{
If the entry \code{typeform_api} exist in your
\code{.Renviron} return that value. Otherwise, raise an error.
}

