% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_responses.R
\name{get_authorization}
\alias{get_authorization}
\title{Get Authorization}
\usage{
get_authorization(api)
}
\arguments{
\item{api}{An authentication key}
}
\description{
A function to get authorization
}
