% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_typeforms.R
\name{get_number_of_forms}
\alias{get_number_of_forms}
\alias{get_forms}
\title{Fetch all available typeforms}
\usage{
get_number_of_forms(api = NULL, search = "", workspace_id = NULL)

get_forms(api = NULL, page = 1, page_size = 10, search = "",
  workspace_id = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api")} is used.}

\item{search}{Returns items that contain the specified string.}

\item{workspace_id}{Retrieve typeforms for the specified workspace.}

\item{page}{The page of results to retrieve. Default 1 is the first page of results.}

\item{page_size}{Number of results to retrieve per page. Default is 10. Maximum is 200.}
}
\value{
A list containing content and the response.
}
\description{
This function returns a two column data frame containing the typeform names and
their associated ids.
}
