% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me.R
\name{get_me}
\alias{get_me}
\title{Retrieve your own account information.}
\usage{
get_me(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api2")} is used.}
}
\description{
Returns alias, email and language.
}
