% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_responses.R
\name{get_response}
\alias{get_response}
\title{Get response}
\usage{
get_response(api, url)
}
\arguments{
\item{api}{An authentication key}

\item{url}{The URL of the site}
}
\description{
A function that allows you to get a response from the
server.
}
