% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Download questionnaire results}
\usage{
get_results(uid, api = NULL, completed = NULL, since = NULL,
  until = NULL, offset = NULL, limit = NULL, order_by = NULL)
}
\arguments{
\item{uid}{The UID (unique identifier) of the typeform you want the results for.}

\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api")} is used.}

\item{completed, }{default \code{NULL}, return all results.
Fetch only completed results (\code{TRUE}), or only not-completed results
(=\code{FALSE}). If \code{NULL} return all results.}

\item{since, }{default \code{NULL}. Fetch only the results after a specific date and
time. If \code{NULL} return all results.}

\item{until, }{default \code{NULL}. Fetch only the results before a specific date and
time. If \code{NULL} return all results.}

\item{offset}{Fetch all results except the first \code{offset}.
i.e. Start listing results from result #\code{offset} onwards.}

\item{limit, }{default \code{NULL}. Fetch only \code{limit} results.
If \code{NULL} return all results.}

\item{order_by}{One of "completed", "date_land_desc", "date_land_incr",
"date_submit_desc", or "date_submit_incr".}
}
\value{
A list containing questions, stats, completed responses,
and uncompleted responses and http status.
}
\description{
Download results for a particular typeform questionnaire.
}
\examples{
\dontrun{
uid = "XXXX"
api = "YYYY"
results = get_results(uid, api)
results$stats
results$questions
results$responses
}
}
\seealso{
https://www.typeform.com/help/data-api/
}
