% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{get_number_of_seats}
\alias{get_number_of_seats}
\alias{get_teams}
\title{Get team members}
\usage{
get_number_of_seats(api = NULL)

get_teams(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api2")} is used.}
}
\description{
Return a tibble containing the current team members. The
total number of seats is attached via an attribute.
}
