% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_typeforms.R
\name{get_typeforms}
\alias{get_typeforms}
\alias{get_all_typeforms}
\title{Fetch all available typeforms}
\usage{
get_typeforms(api = NULL)

get_all_typeforms(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api")} is used.}
}
\value{
A list containing content and the response.
}
\description{
This function returns a two column data frame containing the typeform names and
their associated ids.
}
\examples{
\dontrun{
api = "XXXXX"
tf = get_all_typeforms(api)
tf$content

## Use the system variable
get_all_typeforms()
}
}
