% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{make_new_token}
\alias{make_new_token}
\title{Create an OAuth token}
\usage{
make_new_token()
}
\description{
Create a token object using the client id and secret options set, together
with the defined scopes.
}
