% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_responses.R
\name{post_response}
\alias{post_response}
\title{Post response}
\usage{
post_response(api, url, body = NULL, ...)
}
\arguments{
\item{api}{An authentication key}

\item{url}{The URL of the site}

\item{body}{The infrormation sent out to the server. e.g. the title}

\item{...}{Other arguments}
}
\description{
A function that allows you to post an update to the
server
}
