% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rtypeform_auth_return_code}
\alias{rtypeform_auth_return_code}
\title{Returns the authentication parameter "code" in redirected URLs}
\usage{
rtypeform_auth_return_code(session)
}
\arguments{
\item{session}{A shiny session object}

\item{securityCode}{A random string to check the auth comes form the same origin.}
}
\value{
The auth token in the code URL parameter.
}
\description{
Checks the URL of the Shiny app to get the state and code URL parameters.
}
\concept{shiny auth functions}
\keyword{internal}
