% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rtypeform_set_client_id}
\alias{rtypeform_set_client_id}
\title{Set option for rtypeform client id}
\usage{
rtypeform_set_client_id(id = NULL)
}
\arguments{
\item{id}{A client id from a typeform registered application}
}
\description{
Set option for rtypeform client id
}
