% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rtypeform_set_client_secret}
\alias{rtypeform_set_client_secret}
\title{Set option for rtypeform client secret}
\usage{
rtypeform_set_client_secret(secret = NULL)
}
\arguments{
\item{secret}{A client secret from a typeform registered application}
}
\description{
Set option for rtypeform client secret
}
