% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rtypeform_set_scope}
\alias{rtypeform_set_scope}
\title{Set access scopes for OAuth access}
\usage{
rtypeform_set_scope(scopes = NULL)
}
\arguments{
\item{scopes}{A character vector of scopes}
}
\description{
Define access scopes for an application. See \url{https://developer.typeform.com/get-started/scopes/}
for more information on what scopes are allowed and their permissions.
}
\examples{
rtypeform_set_scope(scopes = c("forms:read"))
}
