% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rtypeform_set_token}
\alias{rtypeform_set_token}
\title{Set the environment variable with access token}
\usage{
rtypeform_set_token(token)
}
\arguments{
\item{token}{Either a string of a personal access token, or an OAuth token returned by make_new_token}
}
\description{
Set the environment variable with access token
}
