% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rtypeform_shiny_get_url}
\alias{rtypeform_shiny_get_url}
\title{Get the Shiny Apps URL.}
\usage{
rtypeform_shiny_get_url(session)
}
\arguments{
\item{session}{The shiny session object.}
}
\value{
The URL of the Shiny App its called from.
}
\description{
Needed for the redirect URL in Auth flow
}
