% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api.R
\name{set_api}
\alias{set_api}
\title{Set typeform api key}
\usage{
set_api(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api")} is used.}
}
\description{
A wrapper around \code{Sys.setenv} for setting the typeform api key for the current session.
}
