% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeformAddin.R
\name{typeformAddin}
\alias{typeformAddin}
\title{Questionnaire selection}
\usage{
typeformAddin(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api} is \code{NULL},
the environment variable \code{Sys.getenv("typeform_api")} is used.}
}
\description{
This rstudio add in is for selecting the appropriate questionnaire from the group of
questionnaires available for a given api key based on matching the questionnaire name
up with the unique id.
}
